refactor(영상 성과 분석): 무한스크롤 관련 타입 분리
src/shared/api/types.ts — PageInfo 추가 (cursor 공통 타입)
src/features/videos/model/types.ts — 신규 생성, VideosResponse 선언
src/features/videos/api/videosApi.ts — 타입 선언 제거, model/types에서 import로 교체
